added length of mouse traveled path in cm

+++ Statistiks.Report/IReportService.cs
@@ -6,6 +6,6 @@ namespace Statistiks.Report
 {
     public interface IReportService
     {
-        void SaveReport(DateTime sessionStart, DateTime sessionEnd, string path, Dictionary<MouseMessage, ulong> mouseUsage, Dictionary<string, ulong> keyboardUsage, Dictionary<string, ulong> windowUsage);
+        void SaveReport(DateTime sessionStart, DateTime sessionEnd, string path, Dictionary<MouseMessage, double> mouseUsage, Dictionary<string, ulong> keyboardUsage, Dictionary<string, ulong> windowUsage);
     }
 }
